简化log模块条件编译的属性设置
